phosphorylation_prediction/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep learning
torch>=1.12.0
transformers>=4.20.0
tokenizers>=0.12.0

# Gradient boosting
xgboost>=1.6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Experiment tracking
wandb>=0.12.0

# Configuration
PyYAML>=6.0
pydantic>=1.9.0

# Utilities
tqdm>=4.64.0
psutil>=5.8.0
joblib>=1.1.0

# Bioinformatics
biopython>=1.79

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.950

# Jupyter (for notebooks)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Optional GPU acceleration
# torch-audio  # Uncomment if needed
# torch-vision  # Uncomment if needed